refactor: 收敛 Agent 展示契约与列车领域投影
