+++ qq-maid-core/src/runtime/respond/chat_flow/mod.rs
@@ -44,7 +44,11 @@ const TOOL_LOOP_AMBIGUITY_PROMPT: &str = "\
 其他工具缺少必要信息时不要调用写工具，直接用自然语言追问并结束本轮回复。\
 字段归位：中文时间词如今天、明天、周四、上午、下午、晚上应进入时间字段或保留在原文，不要当成标题；\
 标题优先表达核心事项，补充目标放 detail，不能把主项和补充说明反转。\
-响应编排：工具执行前如需可见反馈，只能说“我帮你确认一下/试着处理”，不得提前说已完成、已记好或已成功。";
+响应编排：工具执行前如需可见反馈，只能说“我帮你确认一下/试着处理”，不得提前说已完成、已记好或已成功。\
+最终展示契约：如果最终正文确实展示了某次成功只读工具返回的条目，且需要让用户在下一轮继续用“第一条/刚刚那条”等编号指代，\
+请返回严格 JSON 对象 {\"reply\":\"用户可见正文\",\"published_tool_call_ids\":[\"本轮工具结果的 call_id\"]}；\
+数组只列出正文实际展示的结果，call_id 必须来自当前工具轨迹，不要列隐藏、失败或未展示的结果。\
+如果正文没有展示可供后续编号引用的工具条目，直接返回普通自然语言正文，不要猜测 call_id 或伪造展示记录。";
 const GROUP_TOOL_WHITELIST_PROMPT: &str = "\
 群聊工具边界：当前群聊只允许调用本场景配置白名单中的工具；\
 不要声称已经执行未开放的工具，也不要声称写入了未实际修改的持久化状态。";
