+++ qq-maid-core/src/runtime/tools/todo/mod.rs
@@ -56,6 +56,7 @@ use qq_maid_llm::tool::DynTool;
 
 use crate::{runtime::session::SessionStore, storage::notification::NotificationOutboxStore};
 
+pub(crate) use common::LIST_TODOS_TOOL_NAME;
 pub(crate) use complete::CompleteTodoTool;
 pub(crate) use create::CreateTodoTool;
 pub(crate) use delete::DeleteTodoTool;
